피드백 반영
* 다른 생성자가 가장 인자가 많은 생성자를 호출하는 구조로 변경
* 단순숫자를 의미있는 상수화
* label에 작성자 추가하고, 삭제 및 수정은 타인issue가 설정되지 않았을 경우만 가능하도록 변경
* thin service 지향토록 issue Service의 setMilestone 메서드 수정

+++ src/main/java/codesquad/service/LabelService.java
@@ -1,7 +1,9 @@
 package codesquad.service;
 
+import codesquad.UnAuthorizedException;
 import codesquad.domain.Label;
 import codesquad.domain.LabelRepository;
+import codesquad.domain.User;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
@@ -25,6 +27,12 @@ public class LabelService {
                 .orElseThrow(EntityNotFoundException::new);
     }
 
+    public Label findById(User loginUser, long id) {
+        return labelRepository.findById(id)
+                .filter(user -> user.isOwner(loginUser))
+                .orElseThrow(() -> new UnAuthorizedException("You're Unauthorized User"));
+    }
+
     @Transactional
     public Label add(Label label) throws Exception {
         if (labelRepository.findByName(label.getName()).isPresent()) throw new Exception();
@@ -32,14 +40,15 @@ public class LabelService {
     }
 
     @Transactional
-    public void update(long id, Label updatedLabel) {
-        Label label = findById(id);
+    public void update(User loginUser, long id, Label updatedLabel) throws Exception{
+        Label label = findById(loginUser, id);
         label.update(updatedLabel);
     }
 
     @Transactional
-    public void delete(long id) {
-        labelRepository.deleteById(id);
+    public void delete(User loginUser, long id) throws Exception{
+        Label label = findById(loginUser, id);
+        if (label.delete()) labelRepository.deleteById(id);
     }
 
     @Transactional
